// src/app/models/parametersCreate.ts
export class ParametersCreate{
        age: number;
        growth: number;
        clothsize: number;
        chestsize: number;
        waistsize: number;
        thighsize: number;
        footsize: number;
        necksize: number;
        headsize: number;
        tatu: string;
        piercing: string;
        appearancetype :number;
        haircolor: number;
        city: number;
        eyecolor: number;
        gender: number;
        country: number;
        employmenttype: number[] = [];
}